Fix 64-bit pointer truncation in error()/merror() second parameter
The second parameter of error() and merror() was declared as `int`
(32 bits) but callers pass `char *` values through it (e.g. Command
in quickly()). On 64-bit systems, the pointer is truncated, and when
printf tries to use it as a %s string pointer, it crashes. This
caused :q after modification to crash instead of displaying the
"No write since last change" warning.

Changed `int i` to `long i` in both functions, as `long` is 64 bits
on x86-64 and can hold both integer and pointer values, matching the
original K&R varargs behavior on 32-bit VAX.

+++ port/ex_cmds2.c
@@ -77,7 +77,8 @@ eol()
 error(str, i)
 	/* ORIG: register int str; */
 	register char *str;
-	int i;
+	/* ORIG: int i; */
+	long i;
 {
 
 	error0();
